Fix LINQPad sample
Tried running the samples in LINQPad 6.10. This is amazing stuff!

Found one broken that won't compile and run, but the fix is simple. The `Do()` extension method is expecting an `IEnumerable<T>`, but the `new Rectangle()` on line 51 is a single item. I tried to come up with some clever overload or changing behavior based on type checking, but this seemed simpler, and certainly works.